fix: minor type error

+++ src/include/utils.h
@@ -58,10 +58,10 @@ typedef struct ColorData16 {
 } ColorData16;
 
 typedef struct ColorData8 {
-    uint16_t r;
-    uint16_t g;
-    uint16_t b;
-    uint16_t a;
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
+    uint8_t a;
 } ColorData8;
 
 typedef struct RenderData {
